pyproject.toml
[project]
name = "gws-cv-mail-harness"
version = "0.1.0"
description = "Local-first Google Workspace CV mail harness for agent-driven development and validation."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0,<7.0",
  "typer>=0.12,<1.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.11,<0.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
